Refactor OrbitModule/Pdb
* Finally moved the things belonging to pdb into Pdb.cpp (out of
OrbitModule.cpp)
* Deleted a bunch of things that were not used anymore.
