+++ OrbitCore/OrbitModule.cpp
@@ -4,6 +4,8 @@
 
 #include "OrbitModule.h"
 
+#include <absl/strings/str_format.h>
+
 #include <cinttypes>
 #include <string>
 
@@ -11,17 +13,12 @@
 #include "ElfUtils/ElfFile.h"
 #include "OrbitBase/Logging.h"
 #include "Pdb.h"
-#include "ScopeTimer.h"
-#include "absl/strings/str_format.h"
 #include "symbol.pb.h"
 
 #ifndef WIN32
-#include "Capture.h"
 #include "LinuxUtils.h"
-#include "OrbitProcess.h"
 #include "Params.h"
 #include "Path.h"
-#include "ScopeTimer.h"
 #endif
 
 //-----------------------------------------------------------------------------
@@ -34,48 +31,12 @@ Module::Module(const std::string& file_name, uint64_t address_start,
 
   m_FullName = file_name;
   m_Name = Path::GetFileName(file_name);
-  m_Directory = Path::GetDirectory(file_name);
   m_PdbSize = Path::FileSize(file_name);
 
   m_AddressStart = address_start;
   m_AddressEnd = address_end;
 
-  m_PrettyName = m_FullName;
-  m_AddressRange = absl::StrFormat("[%016" PRIx64 " - %016" PRIx64 "]",
-                                   m_AddressStart, m_AddressEnd);
-
   loadable_ = true;  // necessary, because it toggles "Load Symbols" option
-}
-
-//-----------------------------------------------------------------------------
-std::string Module::GetPrettyName() {
-  if (m_PrettyName.empty()) {
-#ifdef WIN32
-    m_PrettyName =
-        absl::StrFormat("%s [%I64x - %I64x] %s\r\n", m_Name.c_str(),
-                        m_AddressStart, m_AddressEnd, m_FullName.c_str());
-    m_AddressRange =
-        absl::StrFormat("[%I64x - %I64x]", m_AddressStart, m_AddressEnd);
-#else
-    m_PrettyName = m_FullName;
-    m_AddressRange =
-        absl::StrFormat("[%016llx - %016llx]", m_AddressStart, m_AddressEnd);
-#endif
-  }
-
-  return m_PrettyName;
-}
-
-//-----------------------------------------------------------------------------
-bool Module::LoadDebugInfo() {
-  CHECK(m_Pdb != nullptr);
-  m_Pdb->SetMainModule(m_AddressStart);
-
-  PRINT_VAR(loadable_);
-  if (!loadable_) return false;
-
-  loaded_ = m_Pdb->LoadDataFromPdb();
-  return loaded_;
 }
 
 //-----------------------------------------------------------------------------
@@ -97,136 +58,5 @@ void Module::LoadSymbols(const ModuleSymbols& module_symbols) {
   }
 
   m_Pdb->ProcessData();
-  m_PdbName = module_symbols.symbols_file_path();
   SetLoaded(true);
-}
-
-//-----------------------------------------------------------------------------
-uint64_t Module::ValidateAddress(uint64_t a_Address) {
-  if (ContainsAddress(a_Address)) return a_Address;
-
-  // Treat input address as RVA
-  uint64_t newAddress = m_AddressStart + a_Address;
-  if (ContainsAddress(newAddress)) return newAddress;
-
-  return 0xbadadd;
-}
-
-//-----------------------------------------------------------------------------
-void Module::SetLoaded(bool value) { loaded_ = value; }
-
-#ifndef WIN32
-
-//-----------------------------------------------------------------------------
-Function* Pdb::FunctionFromName(const std::string& a_Name) {
-  uint64_t hash = StringHash(a_Name);
-  auto iter = m_StringFunctionMap.find(hash);
-  return (iter == m_StringFunctionMap.end()) ? nullptr : iter->second;
-}
-
-//-----------------------------------------------------------------------------
-Pdb::Pdb(uint64_t module_address, uint64_t load_bias, std::string file_name,
-         std::string module_file_name)
-    : m_MainModule(module_address),
-      load_bias_(load_bias),
-      m_FileName(std::move(file_name)),
-      m_LoadedModuleName(std::move(module_file_name)) {
-  m_Name = Path::GetFileName(m_FileName);
-}
-
-//-----------------------------------------------------------------------------
-void Pdb::AddFunction(const std::shared_ptr<Function>& function) {
-  functions_.push_back(function);
-  functions_.back()->SetModulePathAndAddress(GetLoadedModuleName(),
-                                             GetHModule());
-}
-
-//-----------------------------------------------------------------------------
-void Pdb::ProcessData() {
-  std::shared_ptr<Process> process = Capture::GTargetProcess;
-  if (process == nullptr) return;
-
-  SCOPE_TIMER_LOG("ProcessData");
-  ScopeLock lock(process->GetDataMutex());
-
-  for (auto& func : functions_) {
-    func->SetModulePathAndAddress(GetLoadedModuleName(), GetHModule());
-    process->AddFunction(func);
-  }
-
-  SCOPE_TIMER_LOG("Find File and Line info");
-  for (auto& func : functions_) {
-    func->FindFile();
-  }
-
-  PopulateFunctionMap();
-  PopulateStringFunctionMap();
-}
-
-//-----------------------------------------------------------------------------
-void Pdb::PopulateFunctionMap() {
-  SCOPE_TIMER_LOG("Pdb::PopulateFunctionMap");
-  for (auto& function : functions_) {
-    m_FunctionMap.insert(std::make_pair(function->Address(), function.get()));
-  }
-}
-
-//-----------------------------------------------------------------------------
-void Pdb::PopulateStringFunctionMap() {
-  {
-    // SCOPE_TIMER_LOG("Reserving map");
-    m_StringFunctionMap.reserve(static_cast<size_t>(1.5f * functions_.size()));
-  }
-
-  {
-    // SCOPE_TIMER_LOG("Map inserts");
-    for (auto& function : functions_) {
-      m_StringFunctionMap[function->Hash()] = function.get();
-    }
-  }
-}
-
-//-----------------------------------------------------------------------------
-Function* Pdb::GetFunctionFromExactAddress(uint64_t a_Address) {
-  uint64_t function_address = a_Address - GetHModule() + load_bias_;
-  auto it = m_FunctionMap.find(function_address);
-  return (it != m_FunctionMap.end()) ? it->second : nullptr;
-}
-
-//-----------------------------------------------------------------------------
-Function* Pdb::GetFunctionFromProgramCounter(uint64_t a_Address) {
-  if (m_FunctionMap.empty()) {
-    return nullptr;
-  }
-
-  uint64_t relative_address = a_Address - GetHModule() + load_bias_;
-  auto it = m_FunctionMap.upper_bound(relative_address);
-
-  if (it == m_FunctionMap.begin()) {
-    return nullptr;
-  }
-
-  --it;
-  return it->second;
-}
-
-#endif
-
-//-----------------------------------------------------------------------------
-void Pdb::ApplyPreset(const Preset& preset) {
-  SCOPE_TIMER_LOG(absl::StrFormat("Pdb::ApplyPreset - %s", m_Name.c_str()));
-
-  std::string module_name = m_LoadedModuleName;
-  auto it = preset.m_Modules.find(module_name);
-  if (it != preset.m_Modules.end()) {
-    const PresetModule& preset_module = it->second;
-
-    for (uint64_t hash : preset_module.m_FunctionHashes) {
-      auto fit = m_StringFunctionMap.find(hash);
-      if (fit != m_StringFunctionMap.end()) {
-        Function* function = fit->second;
-        function->Select();
-      }
-    }
-  }
 }
