+++ OrbitCore/OrbitModule.h
@@ -21,31 +21,22 @@ struct Module {
   Module(const std::string& file_name, uint64_t address_start,
          uint64_t address_end);
 
-  std::string GetPrettyName();
-  bool LoadDebugInfo();
   void LoadSymbols(const ModuleSymbols& module_symbols);
   bool ContainsAddress(uint64_t a_Address) {
     return m_AddressStart <= a_Address && m_AddressEnd > a_Address;
   }
-  uint64_t ValidateAddress(uint64_t a_Address);
-  void SetLoaded(bool value);
+
+  void SetLoaded(bool value) { loaded_ = value; }
   void SetLoadable(bool value) { loadable_ = value; }
   bool IsLoadable() const { return loadable_; }
   bool IsLoaded() const { return loaded_; }
 
-  std::string m_Name;       // name of the file (without path)
-  std::string m_FullName;   // full filename (including path)
-  std::string m_PdbName;    // full filename of symbols file
-  std::string m_Directory;  // path of the module, without name of the file
-  std::string m_PrettyName;
-  std::string m_AddressRange;
+  std::string m_Name;      // name of the file (without path)
+  std::string m_FullName;  // full filename (including path)
 
   std::string m_DebugSignature;  // gnu build id on linux
-  HMODULE m_ModuleHandle = 0;
   uint64_t m_AddressStart = 0;
   uint64_t m_AddressEnd = 0;
-  uint64_t m_EntryPoint = 0;
-  bool m_Selected = false;
 
   uint64_t m_PdbSize = 0;  // Size in bytes; windows: pdb, linux: module
 
@@ -54,8 +45,6 @@ struct Module {
  private:
   bool loadable_ = false;
   bool loaded_ = false;
-
-  friend class TestRemoteMessages;
 };
 
 #endif  // ORBIT_CORE_ORBIT_MODULE_H_
