+++ OrbitCore/Pdb.h
@@ -10,114 +10,15 @@
 #include <thread>
 #include <vector>
 
-#include "OrbitDbgHelp.h"
 #include "OrbitFunction.h"
 #include "OrbitSession.h"
 
-// TODO: Create a common interface with 2 implementations for Linux and Windows
-// or separate this class into 2 differently named ones. It will hopefully
-// let us remove ifdefs from here and have implementation separated in different
-// .cpp files.
-
-#ifdef _WIN32
-
 class Pdb {
  public:
-  explicit Pdb(const char* pdb_name = "");
   Pdb(uint64_t module_address, uint64_t load_bias, std::string file_name,
       std::string module_file_name);
   Pdb(const Pdb&) = delete;
   Pdb& operator=(const Pdb&) = delete;
-
-  void Init();
-
-  bool LoadDataFromPdb();
-  bool LoadPdbDia();
-  void AddFunction(const std::shared_ptr<Function>& function);
-  void CheckOrbitFunction(Function& a_Function);
-  void PrintFunction(Function& a_Func);
-  void AddArgumentRegister(const std::string& a_Reg,
-                           const std::string& a_Function);
-
-  const std::string& GetName() const { return m_Name; }
-  const std::string& GetFileName() const { return m_FileName; }
-  const std::string& GetLoadedModuleName() const { return m_LoadedModuleName; }
-  const std::vector<std::shared_ptr<Function>>& GetFunctions() {
-    return functions_;
-  }
-  uint64_t GetHModule() const { return m_MainModule; }
-  uint64_t GetLoadBias() const { return load_bias_; }
-
-  GUID GetGuid();
-
-  void SetMainModule(uint64_t a_Module) { m_MainModule = a_Module; }
-  void SetLoadBias(uint64_t load_bias) { load_bias_ = load_bias; }
-
-  void Print() const;
-  void PopulateFunctionMap();
-  void PopulateStringFunctionMap();
-  void Clear();
-  void Reserve();
-  void ApplyPreset(const Preset& preset);
-
-  Function* GetFunctionFromExactAddress(uint64_t a_Address);
-  Function* GetFunctionFromProgramCounter(uint64_t a_Address);
-  bool LineInfoFromAddress(uint64_t a_Address, struct LineInfo& o_LineInfo);
-
-  void SetLoadTime(float a_LoadTime) { m_LastLoadTime = a_LoadTime; }
-  float GetLoadTime() { return m_LastLoadTime; }
-
-  std::string GetCachedName();
-  std::string GetCachedKey();
-
-  bool IsLoading() const { return m_IsLoading; }
-
-  void ProcessData();
-
- protected:
-  // State
-  std::unique_ptr<std::thread> m_LoadingThread;
-  std::atomic<bool> m_FinishedLoading;
-  std::atomic<bool> m_IsLoading;
-  std::atomic<bool> m_IsPopulatingFunctionMap;
-  std::atomic<bool> m_IsPopulatingFunctionStringMap;
-  std::function<void()> m_LoadingCompleteCallback;
-  uint64_t m_MainModule;
-  uint64_t load_bias_ = 0;
-  float m_LastLoadTime;
-  bool m_LoadedFromCache;
-  std::set<std::string> m_ArgumentRegisters;
-  std::map<std::string, std::vector<std::string>> m_RegFunctionsMap;
-
-  // Data
-  std::string m_Name;
-  std::string m_FileName;
-  std::string m_LoadedModuleName;
-  std::vector<std::shared_ptr<Function>> functions_;
-  IMAGEHLP_MODULE64 m_ModuleInfo;
-  std::map<uint64_t, Function*> m_FunctionMap;
-  std::unordered_map<unsigned long long, Function*> m_StringFunctionMap;
-  Timer* m_LoadTimer;
-};
-
-#else
-class Pdb {
- public:
-  explicit Pdb(const char* filename = "") { m_FileName = filename; }
-  Pdb(uint64_t module_address, uint64_t load_bias, std::string file_name,
-      std::string module_file_name);
-  Pdb(const Pdb&) = delete;
-  Pdb& operator=(const Pdb&) = delete;
-
-  void Init() {}
-
-  bool LoadDataFromPdb() {
-    return true;
-  }  // This shouldn't do anything on Linux.
-  bool LoadPdbDia() { return false; }
-  void CheckOrbitFunction(Function&) {}
-  void PrintFunction(Function&) {}
-  void AddArgumentRegister(const std::string&, const std::string&) {}
 
   const std::string& GetName() const { return m_Name; }
   const std::string& GetFileName() const { return m_FileName; }
@@ -129,56 +30,24 @@ class Pdb {
   uint64_t GetHModule() const { return m_MainModule; }
   uint64_t GetLoadBias() const { return load_bias_; }
 
-  GUID GetGuid();
-
-  void SetMainModule(uint64_t a_Module) { m_MainModule = a_Module; }
-  void SetLoadBias(uint64_t load_bias) { load_bias_ = load_bias; }
-
-  void Print() const;
   void PopulateFunctionMap();
   void PopulateStringFunctionMap();
-  void Clear();
-  void Reserve();
   void ApplyPreset(const Preset& preset);
 
   Function* GetFunctionFromExactAddress(uint64_t a_Address);
   Function* GetFunctionFromProgramCounter(uint64_t a_Address);
-  bool LineInfoFromAddress(uint64_t a_Address, struct LineInfo& o_LineInfo);
-  Function* FunctionFromName(const std::string& a_Name);
-
-  void SetLoadTime(float a_LoadTime) { m_LastLoadTime = a_LoadTime; }
-  float GetLoadTime() { return m_LastLoadTime; }
-
-  std::string GetCachedName();
-  std::string GetCachedKey();
-
-  bool IsLoading() const { return m_IsLoading; }
 
   void ProcessData();
 
- protected:
-  // State
-  std::unique_ptr<std::thread> m_LoadingThread;
-  std::atomic<bool> m_FinishedLoading;
-  std::atomic<bool> m_IsLoading;
-  std::function<void()> m_LoadingCompleteCallback;
+ private:
   uint64_t m_MainModule = 0;
   uint64_t load_bias_ = 0;
-  float m_LastLoadTime = 0;
-  std::set<std::string> m_ArgumentRegisters;
-  std::map<std::string, std::vector<std::string>> m_RegFunctionsMap;
-
-  // Data
   std::string m_Name;              // name of the file containing the symbols
   std::string m_FileName;          // full path of file containing the symbols
   std::string m_LoadedModuleName;  // full path of the module
   std::vector<std::shared_ptr<Function>> functions_;
   std::map<uint64_t, Function*> m_FunctionMap;
   std::unordered_map<unsigned long long, Function*> m_StringFunctionMap;
-  Timer* m_LoadTimer = nullptr;
 };
-#endif
-
-extern std::shared_ptr<Pdb> GPdbDbg;
 
 #endif  // ORBIT_CORE_PDB_H_
